[ecore][go] add suffix Z as the timezone is always zero UTC offset
