+++ ecore/ecorefactory_ext.go
@@ -44,7 +44,7 @@ func (factory *ecoreFactoryExt) convertECharToString(dataType EDataType, instanc
 }
 
 const (
-	dateFormat string = "2006-01-02T15:04:05.999"
+	dateFormat string = "2006-01-02T15:04:05.999Z"
 )
 
 func (factory *ecoreFactoryExt) createEDateFromString(dataType EDataType, literalValue string) interface{} {
